#8 Refactor: movie api
