+++ server/routes/movies.js
@@ -3,22 +3,8 @@ const express = require('express');
 const router = express.Router();
 const movieCtrl = require('../controllers/movies');
 
-router.get('/', async (req, res) => {
-  const data = await movieCtrl.getPopularMovies();
-  const data2 = await movieCtrl.getMoviesWithCountry(data);
-  res.send(data2);
-});
-
-router.get('/:movieId', async (req, res) => {
-  const { movieId } = req.params;
-  const data = await movieCtrl.getMoviesDetailsById(movieId);
-  res.send(data);
-});
-
-router.get('/search/:keyword', async (req, res) => {
-  const { keyword } = req.params;
-  const data = await movieCtrl.searchMoviesById(keyword);
-  res.send(data);
-});
+router.get('/', movieCtrl.getBoxoffices);
+router.get('/:movieId', movieCtrl.getMovieDetailById);
+router.get('/search/:keyword', movieCtrl.getSearchMovies);
 
 module.exports = router;
